fix:取消订单时删除

+++ GGG_backend/views.py
@@ -343,14 +343,14 @@ def cancel_order(area, product, openid, job):
         cancel_user = Passenger.objects.filter(name=openid).first()
         if cancel_user.name in passenger_unmatched:
             passenger_unmatched.remove(cancel_user.name)
-        if cancel_user.myorder_id != -1:
-            order_id = cancel_user.myorder_id
-        cancel_user.myorder_id = -1
+        order_id = cancel_user.myorder_id
+        order = Order.objects.filter(id=order_id).first()
         if cancel_user.status < 2:  # 匹配前取消
             cancel_user.status = 0
             cancel_user.save()
+            if order:
+                order.delete()
             return
-        order = Order.objects.filter(id=order_id).first()
         influenced_user = Driver.objects.filter(name=order.mydriver).first()
         if influenced_user.name in driver_matched:
             driver_matched.remove(influenced_user.name)
@@ -358,9 +358,10 @@ def cancel_order(area, product, openid, job):
             match(int(area), int(product), influenced_user.name, "driver")
         if cancel_user.name in passenger_matched:
             passenger_matched.remove(cancel_user.name)
-        influenced_user.myorder_id = -1
-        order.delete()
+        if order:
+            order.delete()
         cancel_user.myorder_id = -1
+        influenced_user.myorder_id = -1
     elif job == "driver":
         cancel_user = Driver.objects.filter(name=openid).first()
         if cancel_user.name in driver_unmatched:
@@ -373,8 +374,7 @@ def cancel_order(area, product, openid, job):
             cancel_user.save()
             return
         order = Order.objects.filter(id=order_id).first()
-        influenced_user = Passenger.objects.filter(
-            name=order.mypassenger).first()
+        influenced_user = Passenger.objects.filter(name=order.mypassenger).first()
         if influenced_user.name in passenger_matched:
             passenger_matched.remove(influenced_user.name)
             passenger_unmatched.insert(0, influenced_user.name)
